requirements.txt
newsdataapi==0.1.8
python-dotenv==1.0.0
pyyaml==6.0.1
pandas==1.5.3
nltk==3.8.1
numpy==1.26.2